requirements.txt
alembic==0.9.5
click==6.7
Flask==0.12.2
Flask-HTTPAuth==3.2.3
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
